feat: comment 생성, 삭제 구현 / user 난수 생성 / logger 생성

+++ src/services/comment.service.js
@@ -1,25 +1,36 @@
+const {
+  BadRequestError,
+  ForbiddenError,
+} = require('../exceptions/customError.js');
+const { logger } = require('../middlewares/logger');
 const CommentRepository = require('../repositories/comment.repository');
 const PinRepository = require('../repositories/pin.repository');
 
 class CommentService {
   constructor() {
-    this.commentRepository = CommentRepository();
-    this.pinRepository = PinRepository();
+    this.commentRepository = new CommentRepository();
+    this.pinRepository = new PinRepository();
   }
-  addComment = async ({ pinId, content }) => {
+  addComment = async ({ userId, pinId, content }) => {
+    logger.info(`CommentService.addComment`);
     const existComment = await this.pinRepository.findByPinId({ pinId });
     if (!existComment) {
-      const err = new Error('게시글 조회에 실패하였습니다.');
-      err.name = '404';
-      throw err;
+      throw new BadRequestError('게시글 조회에 실패하였습니다.');
     }
-    await this.commentRepository.addComment({ pinId, content });
+    await this.commentRepository.addComment({ userId, pinId, content });
 
     return { message: '댓글을 생성하였습니다.' };
   };
 
-  //userId 추가
-  deleteComment = async ({ pinId, commentId }) => {
+  deleteComment = async ({ userId, pinId, commentId }) => {
+    logger.info(`CommentService.deleteComment`);
+    const Comment = await this.commentRepository.findByPinId({ pinId });
+    if (!Comment) {
+      throw new BadRequestError('게시글 조회에 실패하였습니다.');
+    }
+    if (Comment.userId !== userId) {
+      throw new ForbiddenError('권한이 없습니다.');
+    }
     await this.commentRepository.deleteComment({ commentId });
     return { message: '댓글을 삭제하였습니다.' };
   };
